datapath: Fix 32-bit truncation of argument to ODPAT_OUTPUT_CONTROL.
Acked-by: Jesse Gross <jesse@nicira.com>

+++ datapath/actions.c
@@ -360,7 +360,7 @@ error:
 	kfree_skb(skb);
 }
 
-static int output_control(struct datapath *dp, struct sk_buff *skb, u32 arg)
+static int output_control(struct datapath *dp, struct sk_buff *skb, u64 arg)
 {
 	skb = skb_clone(skb, GFP_ATOMIC);
 	if (!skb)
